v1.2 correccion testeo + evaluacion 1

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: juagomez <juagomez@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/17 12:41:23 by juagomez          #+#    #+#             */
-/*   Updated: 2024/08/22 14:42:16 by juagomez         ###   ########.fr       */
+/*   Updated: 2024/08/23 14:44:14 by juagomez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,9 +16,10 @@
 // LIBRERIAS EXTERNAS -----------------------------------
 // ----------------------------------------------------
 
-# include <stdio.h>
 # include <stdarg.h>
 # include <unistd.h>
+// # include <stdlib.h>
+// # include <stdio.h>
 
 // FUNCIONES PRINT  -----------------------------------
 // ----------------------------------------------------
@@ -57,7 +58,7 @@ memoria que se convertirá e imprimirá.
 * @returns int -> número total de caracteres escritos en stdout, 
 incluido el prefijo "0x".
 */
-int	ft_putaddress(size_t ptr_address);
+int	ft_putaddress(unsigned long long ptr_address);
 
 /** 
 * @brief Útil para convertir e imprimir números enteros largos sin signo 
@@ -69,7 +70,7 @@ imprimirá como un número hexadecimal.
 'x' para minúsculas o 'X' para mayúsculas.
 * @returns int -> número total de caracteres escritos 
 */
-int	ft_puthexa(unsigned long num, char specifier_format);
+int	ft_puthexa(unsigned int num, char specifier_format);
 
 /** 
 * @brief Convertir e imprimir un entero largo sin signo como un número decimal
